+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-[Unreleased]
+## [0.0.0-alpha.9] - 2024-04-10
 
 # Added